# backend/docs/swagger.yaml
basePath: /api/
definitions:
...
  api.docSave:
    properties:
      markdown:
        type: string
      message:
        description: commit message; one is made up when empty
        type: string
      path:
        type: string
      sha:
        description: the blob the edit started from; "" creates a new page
        type: string
    type: object
  api.docsIndex:
    properties:
      branch:
        type: string
      canEdit:
        type: boolean
      dir:
        type: string
      error:
        type: string
      nav:
        items:
          $ref: '#/definitions/docsource.NavItem'
        type: array
      repo:
        type: string
    type: object
  api.docsSettings:
    properties:
      branch:
        type: string
      clearToken:
        type: boolean
      dir:
        type: string
      edit:
        type: boolean
      repo:
        description: owner/name, or the repository's GitHub address
        type: string
      token:
        description: '"" keeps the saved token'
        type: string
    type: object
...
  docsource.NavItem:
    properties:
      children:
        items:
          $ref: '#/definitions/docsource.NavItem'
        type: array
      path:
        type: string
      title:
        type: string
    type: object
  docsource.Page:
    properties:
      gitHub:
        description: the file on GitHub
        type: string
      html:
        type: string
      markdown:
        type: string
      path:
        type: string
      sha:
        description: blob being shown; sent back on save to catch edits made elsewhere
        type: string
      title:
        type: string
    type: object
...
  models.Host:
    properties:
      date:
        type: string
      dns:
        type: string
      hw:
        type: string
      id:
        type: integer
      iface:
        type: string
      ip:
        type: string
      known:
        type: integer
      mac:
        type: string
      name:
        type: string
      now:
        type: integer
    type: object
...
info:
  contact:
    url: https://github.com/vipthomps/Gimle
  description: 'Home lab dashboard: network discovery, a map of hosts, bookmarks,
    views and docs'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gimlé API
  version: 1.0.0
paths:
...
  /config/docs:
    post:
      consumes:
      - application/json
      description: An empty Token keeps the saved one; ClearToken removes it. The
        token is never sent back.
      parameters:
      - description: Docs settings
        in: body
        name: docs
        required: true
        schema:
          $ref: '#/definitions/api.docsSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.docsIndex'
      summary: Choose where docs come from
      tags:
      - docs
...
  /docs/file:
    get:
      parameters:
      - description: Path inside the docs folder
        in: query
        name: path
        required: true
        type: string
      responses:
        "200":
          description: OK
      summary: An image from the docs folder
      tags:
      - docs
  /docs/page:
    get:
      description: Rendered HTML plus the Markdown and blob sha an edit starts from
      parameters:
      - description: Page path inside the docs folder, e.g. hosts/nas.md
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docsource.Page'
      summary: One docs page
      tags:
      - docs
    post:
      consumes:
      - application/json
      description: Needs editing turned on and a token that may write to the repository.
        Answers 409 when the page changed on GitHub since Sha.
      parameters:
      - description: Page
        in: body
        name: page
        required: true
        schema:
          $ref: '#/definitions/api.docSave'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/docsource.Page'
      summary: Commit a docs page to GitHub
      tags:
      - docs
  /docs/preview:
    post:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: body
        name: page
        required: true
        schema:
          properties:
            Markdown:
              type: string
            Path:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              HTML:
                type: string
            type: object
      summary: Render Markdown without saving it
      tags:
      - docs
  /edit/{id}/{name}/{known}:
    get:
      description: Update a host's name and optionally toggle its "known" status
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: string
      - description: New name for the host
        in: path
        name: name
        required: true
        type: string
      - description: Pass 'toggle' to flip the known/unknown status
        in: path
        name: known
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Edit host
      tags:
      - hosts
...
  /history/{mac}:
    get:
      description: Retrieve the latest history entries for a specific host by MAC
        address
      parameters:
      - description: MAC address of the host
        in: path
        name: mac
        required: true
        type: string
      - description: Number of history entries to return
        in: query
        name: num
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Host'
            type: array
      summary: Get history by MAC
      tags:
      - history
  /history/{mac}/{date}:
    get:
      description: |-
        Retrieve history for a specific host on a given date
        The date format is flexible and can be:
        - Year only: `2025`
        - Year + month: `2025-09`
        - Full date: `2025-09-06`
        - Full timestamp: `2025-09-06 00:58:26`
      parameters:
      - description: MAC address of the host
        in: path
        name: mac
        required: true
        type: string
      - description: Date filter (supports YYYY, YYYY-MM, YYYY-MM-DD, YYYY-MM-DD HH:mm:ss)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Host'
            type: array
      summary: Get history by date
      tags:
      - history
  /host/{id}:
    get:
      description: Retrieve detailed information about a host by its unique ID
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Host'
      summary: Get host by ID
      tags:
      - hosts
  /host/add/{mac}:
    get:
      description: |-
        Add host by MAC, with optional Name, IP, Hardware
        Returns `models.Host` with this MAC form DB, either just added or existing
      parameters:
      - description: Host MAC
        in: path
        name: mac
        required: true
        type: string
      - description: Name
        in: query
        name: name
        type: string
      - description: IP
        in: query
        name: ip
        type: string
      - description: Hardware
        in: query
        name: hw
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Host'
      summary: Add host manually
      tags:
      - hosts
  /host/del/{id}:
    get:
      description: Remove a host from the database by its unique ID
      parameters:
      - description: Host ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete host
      tags:
      - hosts
...
swagger: "2.0"